feat(security): add option to disallow private IP ranges in webhooks
Closes #2152
