+++ persistence/sql/testhelpers/network.go
@@ -2,11 +2,13 @@ package testhelpers
 
 import (
 	"context"
+	"testing"
+
 	db "github.com/gofrs/uuid"
+
 	courier "github.com/ory/kratos/courier/test"
 	"github.com/ory/kratos/internal/testhelpers"
 	"github.com/ory/kratos/persistence"
-	"testing"
 )
 
 func DefaultNetworkWrapper(t *testing.T, ctx context.Context, p persistence.Persister) (courier.NetworkWrapper, courier.NetworkWrapper) {
